\name{TPXG}
\alias{TPXG}
\docType{package}
\title{
Two Parameter Xgamma & Poisson Xgamma: Regression & Distribution Functions.
}
\description{
The two-parameter Xgamma and Poisson Xgamma distributions are analyzed, covering standard distribution and regression functions, maximum likelihood estimation,
quantile functions, probability density and mass functions, cumulative distribution functions, and random number generation.
}
\details{
\tabular{lll}{
Package: \tab tpxg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2025-03-12\cr
License: \tab GPL-2\cr
}
}
\author{
Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com} and Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Nikolaos Kontemeniotis \email{kontemeniotisn@gmail.com}.
}

\references{
"Wani, M. A., Ahmad, P. B., Para, B. A. and Elah, N. (2023). A new regression model for count data with applications to health care data. International Journal of Data Science and Analytics."

"Sen, S., Chandra, N. and Maiti, S. S. (2018). On properties and applications of a two-parameter XGamma distribution. Journal of Statistical Theory and Applications, 17(4): 674--685."
}



